requirements.txt
attrs
bandit
google-auth
google-cloud-aiplatform
isort
mypy
openai
pre-commit
pyGithub
pytest
pytest-asyncio
pytest-cov
pytest-env
pytest-mock
requests
ruff
setuptools
tox
types-requests
